admin_portal_frontend/requirements.txt
fastapi
uvicorn[standard]
jinja2
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
httpx
python-multipart

